Heso113 & lillap(feature): added auth guard and logic for routing between pages.
