+++ src/app/app-routing.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { LoginFormComponent } from './components/login/login-form/login-form.component';
-import { LoginPageComponent } from './components/login/login-page/login-page.component';
-import { TrainerPageComponent } from './components/dashboard/trainer-page/trainer-page.component';
+import { LoginFormComponent } from './components/forms/login-form/login-form.component';
+import { LoginPageComponent } from './components/pages/login-page/login-page.component';
+import { TrainerPageComponent } from './components/pages/trainer-page/trainer-page.component';
+import { AuthGuard } from './guards/auth.guard';
 
 const routes: Routes = [
   {
@@ -15,7 +16,8 @@ const routes: Routes = [
   },
   {
     path:'trainerPage',
-    component: TrainerPageComponent
+    component: TrainerPageComponent,
+    canActivate: [ AuthGuard ]
   },
   {
     path: '',
