--- src/app/components/login/login-page/login-page.component.html
@@ -1,1 +1,0 @@
-<p>login-page works!</p>
