+++ src/app/components/pages/login-page/login-page.component.html
@@ -1,0 +1,4 @@
+<p>login-page works!</p>
+<div>
+    <button (click)="onLoginClicked()" type="button">Get Started!</button>
+</div>
