+++ src/app/services/auth/auth.service.ts
@@ -6,4 +6,9 @@ import { Injectable } from '@angular/core';
 export class AuthService {
 
   constructor() { }
+
+  login(): string {
+    return localStorage.getItem('trainerName') || '';
+  }
+
 }
